extract data contract validator tests
