+++ dpp/src/tests/data_contract/mod.rs
@@ -1,0 +1,1 @@
+mod validation;
