+++ dpp/src/tests/data_contract/validation/mod.rs
@@ -1,0 +1,1 @@
+mod data_contract_validator_spec;
