+++ dpp/src/tests/mod.rs
@@ -3,5 +3,6 @@ mod identifier_spec;
 
 pub mod utils;
 
+mod data_contract;
 mod identity;
 mod version;
